+++ app/Http/Controllers/PostsController.php
@@ -30,9 +30,11 @@ class PostsController extends Controller
     {
         //
         $post = new \App\Post;
-        $post->user_id = Auth::user()->id;
+        $post->user_id = $request->user_id;
         $post->title = $request->title;
-        $post->content = $request->content;
+        $post->content = $request->post_content;
+        $post->subbreddit_id = $request->subbreddit_id;
+        $post->url = $request->url;
         $post->save();
 
         return $post;
@@ -47,11 +49,15 @@ class PostsController extends Controller
     public function show($id)
     {
         //
+        return \App\Post::find($id);
+
+        /*
         return \App\Post::with([
             'subbreddit',
             'user',
             'comments'
         ])->($id);
+        */
     }
 
     /**
@@ -63,12 +69,15 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
         $post = \App\Post::find($id);
-        $post->user_id = $request->user_id;
-        $post->title = $request->title;
-        $post->content = $request->content;
-        $post->save();
+        if ($post->user_id = \Auth::user()->id) {
+            $post->title = $request->title;
+            $post->content = $request->content;
+            $post->url = $request->url;
+            $post->save();
+        } else {
+            return response ("Unauthorized", 403);
+        }
 
         return $post;
     }
@@ -81,9 +90,14 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        //
         $post = \App\Post::find($id);
-        $post->delete();
+        if ($post->user_id = \Auth::user()->id) {
+            $post->delete();
+        } else {
+            return response ("Unauthorized", 403);
+        }
+
         return $post;
+
     }
 }
